Lots of fixes and changes ( read commit )
added fast heal super buff.
projectiles are no longer deleted if the person who threw them are in the trigger range.
all the correct projectiles are now detected and zaped
zap effect now works correctly.
added a bool for not destroying friendly projectiles.
ults that can be zapped are now reset to 0 if a player throws one and it gets zapped.